Ultimas correcciones de relaciones y query sql

+++ src/main/java/DBD/VideojuegosWeb/entity/Carrito_compras.java
@@ -35,5 +35,5 @@ public class Carrito_compras {
     private Juego juegoid;
 
     @Column(name = "Total")
-    private Double total;
+    private Integer total;
 }
